feat: add crawling settings
